Refactor the RDF action into `enterprise/server`
