+++ enterprise/e2e/auth/hurl/mcp-resources.all.hurl
@@ -44,7 +44,7 @@ jsonpath "$.result.resources[2].uri" == "{{base}}/self/v1/schemas/api/error"
 jsonpath "$.result.resources[2].name" == "Sourcemeta One API Error"
 jsonpath "$.result.resources[2].description" == "The error response format returned by the Sourcemeta One API endpoints"
 jsonpath "$.result.resources[2].mimeType" == "application/schema+json"
-jsonpath "$.result.resources[2].size" == 8155
+jsonpath "$.result.resources[2].size" == 8381
 jsonpath "$.result.resources[2].annotations.priority" == 0
 jsonpath "$.result.resources[3].uri" == "{{base}}/self/v1/schemas/api/list/response"
 jsonpath "$.result.resources[3].name" == "Sourcemeta One List API Response"
@@ -486,7 +486,7 @@ jsonpath "$.result.resources[5].uri" == "{{base}}/self/v1/schemas/api/error"
 jsonpath "$.result.resources[5].name" == "Sourcemeta One API Error"
 jsonpath "$.result.resources[5].description" == "The error response format returned by the Sourcemeta One API endpoints"
 jsonpath "$.result.resources[5].mimeType" == "application/schema+json"
-jsonpath "$.result.resources[5].size" == 8155
+jsonpath "$.result.resources[5].size" == 8381
 jsonpath "$.result.resources[5].annotations.priority" == 0
 jsonpath "$.result.resources[6].uri" == "{{base}}/self/v1/schemas/api/list/response"
 jsonpath "$.result.resources[6].name" == "Sourcemeta One List API Response"
@@ -911,7 +911,7 @@ jsonpath "$.result.resources[3].uri" == "{{base}}/self/v1/schemas/api/error"
 jsonpath "$.result.resources[3].name" == "Sourcemeta One API Error"
 jsonpath "$.result.resources[3].description" == "The error response format returned by the Sourcemeta One API endpoints"
 jsonpath "$.result.resources[3].mimeType" == "application/schema+json"
-jsonpath "$.result.resources[3].size" == 8155
+jsonpath "$.result.resources[3].size" == 8381
 jsonpath "$.result.resources[3].annotations.priority" == 0
 jsonpath "$.result.resources[4].uri" == "{{base}}/self/v1/schemas/api/list/response"
 jsonpath "$.result.resources[4].name" == "Sourcemeta One List API Response"
