+++ enterprise/e2e/html/hurl/mcp-2025-11-25-search_schemas.all.hurl
@@ -396,44 +396,44 @@ jsonpath "$.result.content[2].name" == "My Object"
 jsonpath "$.result.content[2].description" == "An Object Example"
 jsonpath "$.result.content[2].mimeType" == "application/schema+json"
 jsonpath "$.result.content[3].type" == "resource_link"
-jsonpath "$.result.content[3].uri" == "{{base}}/test/full/01"
-jsonpath "$.result.content[3].name" == "Full 01"
-jsonpath "$.result.content[3].description" == "A schema with both a title and a description (01)"
+jsonpath "$.result.content[3].uri" == "{{base}}/test/jsonld/draft7"
+jsonpath "$.result.content[3].name" == "Legacy Dialect"
+jsonpath "$.result.content[3].description" == "A schema on a dialect without annotation support"
 jsonpath "$.result.content[3].mimeType" == "application/schema+json"
 jsonpath "$.result.content[4].type" == "resource_link"
-jsonpath "$.result.content[4].uri" == "{{base}}/test/full/02"
-jsonpath "$.result.content[4].name" == "Full 02"
-jsonpath "$.result.content[4].description" == "A schema with both a title and a description (02)"
+jsonpath "$.result.content[4].uri" == "{{base}}/test/full/01"
+jsonpath "$.result.content[4].name" == "Full 01"
+jsonpath "$.result.content[4].description" == "A schema with both a title and a description (01)"
 jsonpath "$.result.content[4].mimeType" == "application/schema+json"
 jsonpath "$.result.content[5].type" == "resource_link"
-jsonpath "$.result.content[5].uri" == "{{base}}/test/full/03"
-jsonpath "$.result.content[5].name" == "Full 03"
-jsonpath "$.result.content[5].description" == "A schema with both a title and a description (03)"
+jsonpath "$.result.content[5].uri" == "{{base}}/test/full/02"
+jsonpath "$.result.content[5].name" == "Full 02"
+jsonpath "$.result.content[5].description" == "A schema with both a title and a description (02)"
 jsonpath "$.result.content[5].mimeType" == "application/schema+json"
 jsonpath "$.result.content[6].type" == "resource_link"
-jsonpath "$.result.content[6].uri" == "{{base}}/test/full/04"
-jsonpath "$.result.content[6].name" == "Full 04"
-jsonpath "$.result.content[6].description" == "A schema with both a title and a description (04)"
+jsonpath "$.result.content[6].uri" == "{{base}}/test/full/03"
+jsonpath "$.result.content[6].name" == "Full 03"
+jsonpath "$.result.content[6].description" == "A schema with both a title and a description (03)"
 jsonpath "$.result.content[6].mimeType" == "application/schema+json"
 jsonpath "$.result.content[7].type" == "resource_link"
-jsonpath "$.result.content[7].uri" == "{{base}}/test/full/05"
-jsonpath "$.result.content[7].name" == "Full 05"
-jsonpath "$.result.content[7].description" == "A schema with both a title and a description (05)"
+jsonpath "$.result.content[7].uri" == "{{base}}/test/full/04"
+jsonpath "$.result.content[7].name" == "Full 04"
+jsonpath "$.result.content[7].description" == "A schema with both a title and a description (04)"
 jsonpath "$.result.content[7].mimeType" == "application/schema+json"
 jsonpath "$.result.content[8].type" == "resource_link"
-jsonpath "$.result.content[8].uri" == "{{base}}/test/jsonld/conflict"
-jsonpath "$.result.content[8].name" == "Conflicting Mapping"
-jsonpath "$.result.content[8].description" == "A schema whose JSON-LD annotations cannot resolve"
+jsonpath "$.result.content[8].uri" == "{{base}}/test/full/05"
+jsonpath "$.result.content[8].name" == "Full 05"
+jsonpath "$.result.content[8].description" == "A schema with both a title and a description (05)"
 jsonpath "$.result.content[8].mimeType" == "application/schema+json"
 jsonpath "$.result.content[9].type" == "resource_link"
-jsonpath "$.result.content[9].uri" == "{{base}}/test/description-only/01"
-jsonpath "$.result.content[9].name" == "/test/description-only/01"
-jsonpath "$.result.content[9].description" == "A schema with only a description (01)"
+jsonpath "$.result.content[9].uri" == "{{base}}/test/jsonld/conflict"
+jsonpath "$.result.content[9].name" == "Conflicting Mapping"
+jsonpath "$.result.content[9].description" == "A schema whose JSON-LD annotations cannot resolve"
 jsonpath "$.result.content[9].mimeType" == "application/schema+json"
 jsonpath "$.result.content[10].type" == "resource_link"
-jsonpath "$.result.content[10].uri" == "{{base}}/test/description-only/02"
-jsonpath "$.result.content[10].name" == "/test/description-only/02"
-jsonpath "$.result.content[10].description" == "A schema with only a description (02)"
+jsonpath "$.result.content[10].uri" == "{{base}}/test/jsonld/plain"
+jsonpath "$.result.content[10].name" == "Plain"
+jsonpath "$.result.content[10].description" == "A schema without JSON-LD annotations"
 jsonpath "$.result.content[10].mimeType" == "application/schema+json"
 jsonpath "$.result.structuredContent.results" count == 10
 variable "text" jsonpath "$.results" count == 10
@@ -1230,19 +1230,19 @@ jsonpath "$.result.content[2].name" == "My Object"
 jsonpath "$.result.content[2].description" == "An Object Example"
 jsonpath "$.result.content[2].mimeType" == "application/schema+json"
 jsonpath "$.result.content[3].type" == "resource_link"
-jsonpath "$.result.content[3].uri" == "{{base}}/test/full/01"
-jsonpath "$.result.content[3].name" == "Full 01"
-jsonpath "$.result.content[3].description" == "A schema with both a title and a description (01)"
+jsonpath "$.result.content[3].uri" == "{{base}}/test/jsonld/draft7"
+jsonpath "$.result.content[3].name" == "Legacy Dialect"
+jsonpath "$.result.content[3].description" == "A schema on a dialect without annotation support"
 jsonpath "$.result.content[3].mimeType" == "application/schema+json"
 jsonpath "$.result.content[4].type" == "resource_link"
-jsonpath "$.result.content[4].uri" == "{{base}}/test/full/02"
-jsonpath "$.result.content[4].name" == "Full 02"
-jsonpath "$.result.content[4].description" == "A schema with both a title and a description (02)"
+jsonpath "$.result.content[4].uri" == "{{base}}/test/full/01"
+jsonpath "$.result.content[4].name" == "Full 01"
+jsonpath "$.result.content[4].description" == "A schema with both a title and a description (01)"
 jsonpath "$.result.content[4].mimeType" == "application/schema+json"
 jsonpath "$.result.content[5].type" == "resource_link"
-jsonpath "$.result.content[5].uri" == "{{base}}/test/full/03"
-jsonpath "$.result.content[5].name" == "Full 03"
-jsonpath "$.result.content[5].description" == "A schema with both a title and a description (03)"
+jsonpath "$.result.content[5].uri" == "{{base}}/test/full/02"
+jsonpath "$.result.content[5].name" == "Full 02"
+jsonpath "$.result.content[5].description" == "A schema with both a title and a description (02)"
 jsonpath "$.result.content[5].mimeType" == "application/schema+json"
 jsonpath "$.result.structuredContent.results" count == 5
 variable "text" jsonpath "$.results" count == 5
