+++ enterprise/e2e/html/hurl/schemas-rdf.all.hurl
@@ -570,7 +570,7 @@ jsonpath "$.valid" == true
 
 # Availability follows the evaluate flag whether or not the schema
 # carries JSON-LD keywords
-POST {{base}}/self/v1/api/schemas/rdf/test/no-blaze/jsonld
+POST {{base}}/self/v1/api/schemas/rdf/no-blaze/jsonld
 Content-Type: application/json
 {
   "instance": {}
@@ -613,7 +613,7 @@ header "X-Frame-Options" not exists
 header "Date" matches /^(Mon|Tue|Wed|Thu|Fri|Sat|Sun), (0[1-9]|[12][0-9]|3[01]) (Jan|Feb|Mar|Apr|May|Jun|Jul|Aug|Sep|Oct|Nov|Dec) [0-9]{4} ([01][0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9] GMT$/
 jsonpath "$.valid" == true
 
-POST {{base}}/self/v1/api/schemas/rdf/test/no-blaze/string
+POST {{base}}/self/v1/api/schemas/rdf/no-blaze/string
 Content-Type: application/json
 {
   "instance": "anything"
