+++ enterprise/e2e/html/schemas/test/jsonld/draft7.json
@@ -1,6 +1,5 @@
 {
   "$schema": "http://json-schema.org/draft-07/schema#",
-  "$id": "https://example.com/jsonld/draft7",
   "title": "Legacy Dialect",
   "description": "A schema on a dialect without annotation support",
   "type": "object",
