+++ src/actions/CMakeLists.txt
@@ -23,7 +23,6 @@ target_link_libraries(sourcemeta_one_actions PRIVATE sourcemeta::core::uri)
 target_link_libraries(sourcemeta_one_actions PUBLIC sourcemeta::core::jsonrpc)
 target_link_libraries(sourcemeta_one_actions PRIVATE sourcemeta::core::mcp)
 target_link_libraries(sourcemeta_one_actions PRIVATE sourcemeta::core::jsonpointer)
-target_link_libraries(sourcemeta_one_actions PRIVATE sourcemeta::core::jsonld)
 target_link_libraries(sourcemeta_one_actions PRIVATE sourcemeta::core::io)
 target_link_libraries(sourcemeta_one_actions PRIVATE sourcemeta::core::text)
 target_link_libraries(sourcemeta_one_actions PRIVATE sourcemeta::core::time)
